Rename to IVenueUploadProcessor to IFileUploadProcessor
There's a lot of logic here that we will need for Courses,
Apprenticeships etc. This creates a single type for all Data
Management-related things.
