+++ src/Dfc.CourseDirectory.Core/DataManagement/FileUploadProcessor.Venues.cs
@@ -1,50 +1,16 @@
 ﻿using System;
-using System.Buffers;
 using System.IO;
-using System.Linq;
 using System.Reactive.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure.Storage.Blobs;
-using Dfc.CourseDirectory.Core.DataStore.Sql;
 using Dfc.CourseDirectory.Core.DataStore.Sql.Queries;
 using Dfc.CourseDirectory.Core.Models;
 
 namespace Dfc.CourseDirectory.Core.DataManagement
 {
-    public class VenueUploadProcessor : IVenueUploadProcessor
+    public partial class FileUploadProcessor
     {
-        private static readonly byte[] _bom = new byte[] { 0xEF, 0xBB, 0xBF };
-
-        private static bool _containerIsKnownToExist;
-
-        private readonly ISqlQueryDispatcherFactory _sqlQueryDispatcherFactory;
-        private readonly BlobContainerClient _blobContainerClient;
-        private readonly IClock _clock;
-        private readonly TimeSpan _pollInterval;
-
-        public VenueUploadProcessor(
-            ISqlQueryDispatcherFactory sqlQueryDispatcherFactory,
-            BlobServiceClient blobServiceClient,
-            IClock clock) : this(
-                sqlQueryDispatcherFactory, blobServiceClient, clock, TimeSpan.FromMilliseconds(500))
-        {
-        }
-
-        // Constructor for overriding pollInterval for testing
-        internal VenueUploadProcessor(
-            ISqlQueryDispatcherFactory sqlQueryDispatcherFactory,
-            BlobServiceClient blobServiceClient,
-            IClock clock,
-            TimeSpan pollInterval)
-        {
-            _sqlQueryDispatcherFactory = sqlQueryDispatcherFactory;
-            _blobContainerClient = blobServiceClient.GetBlobContainerClient(Constants.ContainerName);
-            _clock = clock;
-            _pollInterval = pollInterval;
-        }
-
-        public IObservable<UploadStatus> GetUploadStatusUpdates(Guid venueUploadId) =>
+        public IObservable<UploadStatus> GetVenueUploadStatusUpdates(Guid venueUploadId) =>
             Observable.Create<UploadStatus>(async (observer, cancellationToken) =>
             {
                 // The IsolationLevel override here is important - our default Snapshot would never see data changes
@@ -76,7 +42,7 @@ namespace Dfc.CourseDirectory.Core.DataManagement
                 }
             }).DistinctUntilChanged();
 
-        public async Task ProcessFile(Guid venueUploadId, Stream stream)
+        public async Task ProcessVenueFile(Guid venueUploadId, Stream stream)
         {
             using (var dispatcher = _sqlQueryDispatcherFactory.CreateDispatcher())
             {
@@ -107,7 +73,7 @@ namespace Dfc.CourseDirectory.Core.DataManagement
             }
         }
 
-        public async Task<SaveFileResult> SaveFile(Guid providerId, Stream stream, UserInfo uploadedBy)
+        public async Task<SaveFileResult> SaveVenueFile(Guid providerId, Stream stream, UserInfo uploadedBy)
         {
             if (!stream.CanRead)
             {
@@ -192,78 +158,9 @@ namespace Dfc.CourseDirectory.Core.DataManagement
             }
         }
 
-        public Task WaitForProcessingToComplete(Guid venueUploadId, CancellationToken cancellationToken) =>
-            GetUploadStatusUpdates(venueUploadId)
+        public Task WaitForVenueProcessingToComplete(Guid venueUploadId, CancellationToken cancellationToken) =>
+            GetVenueUploadStatusUpdates(venueUploadId)
                 .TakeUntil(status => status == UploadStatus.Processed || status.IsTerminal())
                 .ForEachAsync(_ => { }, cancellationToken);
-
-        private static async Task<bool> FileIsEmpty(Stream stream)
-        {
-            if (stream.Length == 0)
-            {
-                return true;
-            }
-
-            // The file could be empty except for BOM. Check for that too.
-            if (stream.Length == 3)
-            {
-                var buffer = new byte[3];
-
-                await stream.ReadAsync(buffer, 0, 3);
-                stream.Seek(-3, SeekOrigin.Current);
-
-                if (buffer.SequenceEqual(_bom))
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        private static async Task<bool> LooksLikeCsv(Stream stream)
-        {
-            // Check the file looks like a CSV. CsvReader will read whatever data it's given until it finds column and
-            // row separators, even if the file is binary, so we can't use it and get meaningful errors out.
-            // The check here tries to read a line of data, up to 512 bytes. (512 bytes is ample to fit a row of our
-            // headers + some custom ones). If we can successfully read a line and its contents are valid ASCII
-            // then that's a good enough signal.
-
-            const int readBufferSize = 512;
-
-            var buffer = ArrayPool<byte>.Shared.Rent(readBufferSize);
-
-            try
-            {
-                var bytesRead = await stream.ReadAsync(buffer, 0, readBufferSize);
-
-                // Restore the Stream's position
-                stream.Seek(-bytesRead, SeekOrigin.Current);
-
-                for (int i = 0; i < bytesRead; i++)
-                {
-                    byte c = buffer[i];
-
-                    if (c > 127)
-                    {
-                        // Outside of ASCII
-                        return false;
-                    }
-
-                    if (c == '\n' && i > 1)
-                    {
-                        // We've hit the end of the line without errors
-                        return true;
-                    }
-                }
-
-                // We've finished reading the whole buffer and not found an EOL - probably not a CSV
-                return false;
-            }
-            finally
-            {
-                ArrayPool<byte>.Shared.Return(buffer);
-            }
-        }
     }
 }
