+++ src/Dfc.CourseDirectory.WebV2/ServiceCollectionExtensions.cs
@@ -169,7 +169,7 @@ namespace Dfc.CourseDirectory.WebV2
             services.AddTransient<Features.TLevels.AddTLevel.Details.CommandValidator>();
             services.AddTransient<Features.TLevels.ViewAndEditTLevel.EditTLevelJourneyModelFactory>();
             services.AddSingleton<IRegionCache, RegionCache>();
-            services.AddTransient<IVenueUploadProcessor, VenueUploadProcessor>();
+            services.AddTransient<IFileUploadProcessor, FileUploadProcessor>();
 
             if (!environment.IsTesting())
             {
